20231101 add isFunction
