+++ orderTest/orderTest.csproj
@@ -48,6 +48,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="addons\IsFunctions.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="commands\MainCommand.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
